feat: move chartdata
