+++ src/Components/GrowthChart/GrowthChart.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { GrowthChartBuilder } from "./GrowthChartBuilder";
-import { chartData } from "../../DataSets/ChartData";
+import { chartData } from "../../DataSets/newWhoStandardDataSets/ZScores/ChartDataZscores";
 import { useRangeTimePeriode } from './useRangeTimePeriode';
 import { ChartCodes } from "../../types/chartDataTypes";
 
 export const GrowthChart = () => {
     const categoryDataSets = chartData["Weight-for-age GIRLS"];
-    const dataSetEntry = categoryDataSets.datasets[ChartCodes.wfa_g_0_5_z];
+    const dataSetEntry = categoryDataSets.datasets[ChartCodes.wfa_g_0_5_y_z];
     
     const dataSetValues = dataSetEntry.datasetValues;
     const dataSetMetadata = dataSetEntry.metadata;
