--- src/DataSets/WhoStandardDataSets/index.ts
@@ -1,2 +1,0 @@
-export { default as wfa_g_0_5_z } from './wfa-girls-0-5-zscores.json';
-
